+++ tests/integration.rs
@@ -1,0 +1,2 @@
+use elevator::*;
+
